feat(parser): use path-based JSON Schema resolver
