+++ internal/oas/parser/ref.go
@@ -7,27 +7,9 @@ import (
 
 	"github.com/ogen-go/ogen"
 	"github.com/ogen-go/ogen/internal/oas"
-	"github.com/ogen-go/ogen/jsonschema"
 )
 
 type resolveCtx map[string]struct{}
-
-type componentsResolver map[string]*ogen.Schema
-
-func (c componentsResolver) ResolveReference(ref string) (*jsonschema.RawSchema, error) {
-	const prefix = "#/components/schemas/"
-	if !strings.HasPrefix(ref, prefix) {
-		return nil, errors.Errorf("invalid schema reference %q", ref)
-	}
-
-	name := strings.TrimPrefix(ref, prefix)
-	s, ok := c[name]
-	if !ok {
-		return nil, errors.New("schema not found")
-	}
-
-	return s.ToJSONSchema(), nil
-}
 
 func (p *parser) resolveRequestBody(ref string, ctx resolveCtx) (*oas.RequestBody, error) {
 	const prefix = "#/components/requestBodies/"
